requirements.txt
aiogram>=3.27.0
python-dotenv>=1.2.2
pytest>=9.0.3
pytest-asyncio>=1.3.0
